Erorr handling, example reorganization
