--- com-api-gen/Cargo.toml
@@ -1,20 +1,0 @@
-[workspace]
-resolver = "3"
-members = [
-    "com-api-gen",
-    "com-api-gen-iceoryx",
-    "com-api-gen-lola", 
-]
-
-[workspace.package]
-name = "com-api-gen"
-version = "0.1.0"
-edition = "2024"
-publish = ["common"]
-license = "Apache-2.0"
-
-[workspace.dependencies]
-com-api = { path = "../com-api/com-api" }
-com-api-gen-iceoryx = { path = "com-api-gen-iceoryx" }
-com-api-gen-lola = { path = "com-api-gen-lola" }
-iceoryx2 = { git = "https://github.com/eclipse-iceoryx/iceoryx2.git", version = "0.6.1" }
