+++ com-api/com-api-runtime-lola/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2024"
 publish = ["common"]
 
 [dependencies]
-com-api-concept = { path = "../com-api-concept", features = ["lola"] }
+com-api-concept = { workspace = true, features = ["lola"] }
 
 [dev-dependencies]
 futures = "0.3"
