+++ com-api/examples/basic-consumer-producer/main.rs
@@ -1,0 +1,60 @@
+// Copyright (c) 2025 Contributors to the Eclipse Foundation
+//
+// See the NOTICE file(s) distributed with this work for additional
+// information regarding copyright ownership.
+//
+// This program and the accompanying materials are made available under the
+// terms of the Apache License Version 2.0 which is available at
+// <https://www.apache.org/licenses/LICENSE-2.0>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+use com_api::prelude::*;
+use com_api_gen::prelude::*;
+
+fn main() {
+    let runtime_builder = RuntimeBuilderImpl::new();
+        let runtime = runtime_builder.build().unwrap();
+        let producer_builder = runtime.producer_builder::<VehicleInterface>(InstanceSpecifier {
+            specifier: "My/Funk/ServiceName".to_string(),
+        });
+        let producer = producer_builder.build().unwrap();
+        let offered_producer = producer.offer().unwrap();
+
+        // Business logic
+        let uninit_sample = offered_producer.left_tire.allocate().unwrap();
+        let sample = uninit_sample.write(Tire {});
+        sample.send().unwrap();
+
+        // Create service discovery
+        let consumer_discovery = runtime.find_service::<VehicleInterface>(InstanceSpecifier {
+            specifier: "My/Funk/ServiceName".to_string(),
+        });
+        let available_service_instances = consumer_discovery.get_available_instances().unwrap();
+
+        // Create consumer from first discovered service
+        let consumer_builder = available_service_instances
+            .into_iter()
+            .find(|desc| desc.get_instance_id() == 42)
+            .unwrap();
+        let consumer = consumer_builder.build().unwrap();
+
+        // Subscribe to one event
+        let subscribed = consumer.left_tire.subscribe(3).unwrap();
+
+        // Create sample buffer to be used during receive
+        let mut sample_buf = SampleContainer::new();
+        for _ in 0..10 {
+            let uninit_sample = offered_producer.left_tire.allocate().unwrap();
+            let sample = uninit_sample.write(Tire {});
+            sample.send().unwrap();
+            match subscribed.try_receive(&mut sample_buf, 1) {
+                Ok(0) => panic!("No sample received"),
+                Ok(x) => {
+                    let sample = sample_buf.pop_front().unwrap();
+                    println!("{} sample received: sample[0] = {:?}", x, *sample)
+                }
+                Err(e) => panic!("{:?}", e),
+            }
+        }
+}
