+++ com-api/examples/com-api-gen/com-api-gen-iceoryx/Cargo.toml
@@ -6,5 +6,5 @@ publish = ["common"]
 
 [dependencies]
 iceoryx2 = { git = "https://github.com/eclipse-iceoryx/iceoryx2.git", version = "0.6.1" }
-com-api-runtime-iceoryx = { path = "../../com-api/com-api-runtime-iceoryx" }
-com-api = { path = "../../com-api/com-api", features = ["iceoryx"] }
+com-api-runtime-iceoryx = { path = "../../../com-api-runtime-iceoryx" }
+com-api = { path = "../../../com-api", features = ["iceoryx"] }
