+++ com-api/examples/com-api-gen/src/lib.rs
